refactor: validate 관련 함수 폴더 생성 및 함수명 수정
